+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
 # 1.0.2
-- Added flag to perform a forced remove with remove command.
+- Added flag to skip optional files at insert and install commands
+- Added flag to perform a forced remove with remove command
 - Changed how version number was printed with version command
-- Fixed a problem with luabundle command
+- Fixed a problem for different bundle files in luabundle command
 - Fixed a bug with the new Mercury packages standard
 - Added ansicolor to support text color printing in old Windows versions
 - Fixed a bug related to temporary files not being deleted